requirements.txt
fastapi
uvicorn
pydantic
scikit-learn
joblib
pandas
langgraph
langchain-openai
python-dotenv
langchain
streamlit
requests
google-generativeai